third_party/blink/web_tests/fast/replaced/no-focus-ring-embed-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
CONSOLE MESSAGE: Blink Test Plugin: MouseDown at (5,5)
CONSOLE MESSAGE: Blink Test Plugin: MouseUp at (5,5)

This test requires DumpRenderTree. Click the orange border, a focus ring should not appear.
EMBED Event: focus PASS
